backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.28.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.6.2
boto3==1.34.125
botocore==1.34.128
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.4
decorator==5.1.1
deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastembed==0.2.7
filelock==3.15.1
flasgger==0.9.7.1
flask==2.1.3
flask-cors==3.0.10
flask-migrate==3.1.0
flask-sqlalchemy==2.5.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.13
langchain-text-splitters==0.2.1
langfuse==2.35.2
langgraph==0.1.4
langsmith==0.1.82
llama-index==0.10.44
llama-index-agent-openai==0.1.7
llama-index-callbacks-langfuse==0.1.4
llama-index-cli==0.1.11
llama-index-core==0.10.44
llama-index-embeddings-fastembed==0.1.4
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-anthropic==0.1.13
llama-index-llms-gemini==0.1.10
llama-index-llms-openai==0.1.21
llama-index-multi-modal-llms-openai==0.1.4
llama-index-packs-resume-screener==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.13
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.0
llamaindex-py-client==0.1.19
loguru==0.7.2
mako==1.3.5
markupsafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
mypy==1.10.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.0
openai==1.35.1
orjson==3.10.5
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
pip==24.0
pluggy==1.5.0
prompt-toolkit==3.0.46
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.7.4
pydantic-core==2.18.4
pyexecjs==1.5.1
pygments==2.18.0
pymupdf==1.24.4
pymupdfb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
pyreadline3==3.4.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
sentry-sdk==1.7.2
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlalchemy==2.0.28
stack-data==0.6.3
starlette==0.36.3
striprtf==0.0.26
sympy==1.12.1
tenacity==8.4.2
tiktoken==0.7.0
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.2
traitlets==5.14.3
treelib==1.7.0
types-colorama==0.4.15.20240311
types-flask-cors==4.0.0.20240523
types-psycopg2==2.9.21.20240417
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uv==0.2.18
uvicorn==0.28.0
voluptuous==0.13.1
voluptuous-stubs==0.1.1
wcwidth==0.2.13
werkzeug==2.3.7
wheel==0.43.0
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.4
